Split function interface out of converter and simplify stateful tests
Move _FunctionInterface and the state dtype constraints into
function_interface.py. State-only functions now return a fixed
state_update_token output instead of re-exposing every updated state.
Share conversion/comparison helpers between the ordinary, symbolic and
stateful test harnesses, and rename extra_nonstate_steps to
subsequent_inputs.
