+++ stablehlo_coreml/converter.py
@@ -1,4 +1,3 @@
-from dataclasses import dataclass
 from functools import partial, reduce
 
 import numpy as np
@@ -6,7 +5,7 @@ from coremltools import _logger as logger
 from coremltools.converters.mil import mil
 from coremltools.converters.mil._deployment_compatibility import AvailableTarget
 from coremltools.converters.mil.mil import Builder as mb
-from coremltools.converters.mil.mil import Function, Placeholder, Program, Symbol, Var, types
+from coremltools.converters.mil.mil import Function, Program, types
 from coremltools.converters.mil.mil.ops.defs._utils import (
     promote_input_dtypes,
 )
@@ -83,6 +82,7 @@ from jaxlib.mlir.dialects.stablehlo import (
     XorOp,
 )
 
+from .function_interface import _FunctionInterface
 from .ops_register import StableHloOpsRegistry, register_composite_op, register_stablehlo_op
 from .padding import pad_with_cast
 from .passes.utils import register_optimizations
@@ -91,9 +91,7 @@ from .sort_utils import match_sort
 from .state import (
     FunctionStateMapping,
     StateMapping,
-    StateSpec,
     normalize_function_state_maps,
-    preferred_argument_name,
     resolve_state_map,
 )
 from .translation_context import DYNAMIC_DIM_SENTINEL, TranslationContext
@@ -114,10 +112,6 @@ from .utils import (
     zero_tensor,
 )
 
-# Core ML can only serialize state tensors as fp16. fp32 HLO state is
-# stored as fp16 and cast to/from the computation dtype around read/write.
-_STATE_VALUE_DTYPES = frozenset({types.fp16, types.fp32})
-
 
 def convert(
     module,
@@ -133,10 +127,10 @@ def convert(
     modules.
 
     State arguments may be identified by index or by their JAX/MLIR name.
-    Updated outputs are omitted from the function results, except when
-    every result is a state write: Core ML requires at least one output,
-    so the updated values are then kept as results. State is stored as
-    fp16; fp32 values are cast around the read/write.
+    Updated outputs are omitted from the function results. When every result
+    is a state write, a one-element token is returned because Core ML requires
+    at least one tensor output. State is stored as fp16; fp32 values are cast
+    around the read/write.
     """
     if minimum_deployment_target < AvailableTarget.iOS18:
         raise ValueError("Converting to <iOS18 is not supported")
@@ -172,145 +166,6 @@ def _normalize_module(module: ir.Module) -> None:
         context=module.context,
     )
     pm.run(module.operation)
-
-
-@dataclass
-class _FunctionInterface:
-    inputs: dict[str, Placeholder]
-    argument_names: dict[int, str]
-    state_specs: dict[str, StateSpec]
-    state_compute_dtypes: dict[str, object]
-
-    @classmethod
-    def from_hlo(
-        cls,
-        hlo_func: FuncOp,
-        state_map: dict[int, StateSpec],
-    ) -> "_FunctionInterface":
-        inputs = {}
-        state_specs: dict[str, StateSpec] = {}
-        state_compute_dtypes: dict[str, object] = {}
-        argument_names: dict[int, str] = {}
-        used_input_names: set[str] = set()
-        explicit_state_names = {
-            spec.name: in_idx
-            for in_idx, spec in state_map.items()
-            if spec.name is not None
-        }
-        sym_counter = 0
-
-        for in_idx, arg in enumerate(hlo_func.arguments):
-            shape = arg.type.shape
-            context_name = arg.get_name()
-            state_spec = state_map.get(in_idx)
-            explicit_name = state_spec.name if state_spec is not None else None
-            name = explicit_name or preferred_argument_name(arg)
-            if explicit_name is not None:
-                if name in used_input_names:
-                    raise ValueError(
-                        f"Core ML state name {name!r} conflicts with another function input"
-                    )
-            else:
-                if name in used_input_names or (
-                    name in explicit_state_names and explicit_state_names[name] != in_idx
-                ):
-                    name = context_name.lstrip("%")
-                while name in used_input_names or (
-                    name in explicit_state_names and explicit_state_names[name] != in_idx
-                ):
-                    name = f"{name}_input"
-            used_input_names.add(name)
-            argument_names[in_idx] = name
-
-            # Reject dynamic state before constructing MIL Symbols — Symbol
-            # names are process-global and would leak if we raise afterwards.
-            if state_spec is not None and any(d == DYNAMIC_DIM_SENTINEL for d in shape):
-                raise ValueError(f"State input {name} must have a static shape, got {shape}")
-            if shape == []:
-                shape = [1]
-            else:
-                new_shape = []
-                for dim in shape:
-                    if dim == DYNAMIC_DIM_SENTINEL:
-                        new_shape.append(Symbol(f"dim_{sym_counter}"))
-                        sym_counter += 1
-                    else:
-                        new_shape.append(dim)
-                shape = new_shape
-
-            dtype = get_mil_type_from_ir(arg.type.element_type)
-            if state_spec is None:
-                inputs[name] = mb.placeholder(shape=shape, dtype=dtype)
-                continue
-            if dtype not in _STATE_VALUE_DTYPES:
-                raise ValueError(
-                    f"State input {name} has dtype {dtype_str(dtype)}, "
-                    "but Core ML states must be floating point (stored as fp16)"
-                )
-            inputs[name] = mb.state_tensor_placeholder(shape=shape, dtype=types.fp16)
-            state_specs[name] = state_spec
-            state_compute_dtypes[name] = dtype
-
-        return cls(
-            inputs=inputs,
-            argument_names=argument_names,
-            state_specs=state_specs,
-            state_compute_dtypes=state_compute_dtypes,
-        )
-
-    def bind_arguments(
-        self,
-        context: TranslationContext,
-        hlo_func: FuncOp,
-        ssa_func: Function,
-    ) -> dict[str, Var]:
-        state_vars = {}
-        for in_idx, arg in enumerate(hlo_func.arguments):
-            name = self.argument_names[in_idx]
-            var = ssa_func.inputs[name]
-            if name in self.state_specs:
-                state_vars[name] = var
-                var = mb.read_state(input=var)
-                compute_dtype = self.state_compute_dtypes[name]
-                if compute_dtype != types.fp16:
-                    var = mb.cast(x=var, dtype=dtype_str(compute_dtype))
-            context.add_variable(arg.get_name(), var)
-        return state_vars
-
-    def finalize_outputs(
-        self,
-        outputs: list[Var],
-        state_vars: dict[str, Var],
-    ) -> list[Var]:
-        consumed = set()
-        updated_states = []
-        for name, spec in self.state_specs.items():
-            if spec.output is None:
-                continue
-            out_idx = spec.output
-            if not 0 <= out_idx < len(outputs):
-                raise ValueError(
-                    f"State output index {out_idx} is out of range for a function "
-                    f"with {len(outputs)} outputs"
-                )
-            value = outputs[out_idx]
-            if self.state_compute_dtypes[name] != types.fp16:
-                value = mb.cast(x=value, dtype="fp16")
-            updated = mb.coreml_update_state(state=state_vars[name], value=value)
-            updated_states.append((updated, name))
-            consumed.add(out_idx)
-
-        final_outputs = [out for i, out in enumerate(outputs) if i not in consumed]
-        if final_outputs or not updated_states:
-            return final_outputs
-
-        # Core ML requires at least one output, so retain state-only results.
-        for updated, name in updated_states:
-            compute_dtype = self.state_compute_dtypes[name]
-            if compute_dtype != types.fp16:
-                updated = mb.cast(x=updated, dtype=dtype_str(compute_dtype))
-            final_outputs.append(updated)
-        return final_outputs
 
 
 class StableHloConverter(metaclass=StableHloOpsRegistry):
